fix: Delete Decoration.swift
